rbac 权限拦截开发

+++ rbac/rbac-my-permission/src/main/java/tech/edwardvan/rbacmypermission/service/SysAclModuleService.java
@@ -29,6 +29,9 @@ public class SysAclModuleService {
     @Autowired
     private SysAclModuleMapper sysAclModuleMapper;
 
+    @Autowired
+    private SysAclMapper sysAclMapper;
+
     public void save(AclModuleParam param) {
         ValidatorUtil.check(param);
         if (checkExist(param.getParentId(), param.getName(), param.getId())) {
@@ -71,10 +74,9 @@ public class SysAclModuleService {
         if (sysAclModuleMapper.countByParentId(aclModule.getId()) > 0) {
             throw new ParamException("当前模块下面有子模块，无法删除");
         }
-        //todo
-//        if (sysAclMapper.countByAclModuleId(aclModule.getId()) > 0) {
-//            throw new ParamException("当前模块下面有用户，无法删除");
-//        }
+        if (sysAclMapper.countByAclModuleId(aclModule.getId()) > 0) {
+            throw new ParamException("当前模块下面有权限，无法删除");
+        }
         sysAclModuleMapper.deleteByPrimaryKey(aclModuleId);
     }
